Add username and token caching, and /autocomplete

+++ app/database.py
@@ -29,7 +29,7 @@ from asyncpg.exceptions import InvalidAuthorizationSpecificationError, PostgresC
 from contextlib import asynccontextmanager
 from datetime import date, datetime
 from sys import stderr
-from typing import TYPE_CHECKING, Literal, TypeAlias
+from typing import TYPE_CHECKING, Literal, TypeAlias, overload
 from uuid import UUID
 from .errors import *
 
@@ -40,9 +40,62 @@ if TYPE_CHECKING:
 ConnCtxMgr: TypeAlias = asyncpg.pool.PoolConnectionProxy
 OptConn: TypeAlias = ConnCtxMgr | None
 
+class UserCache:
+    usernames: dict[str, UUID | None]
+    tokens: dict[UUID, str]
+
+    def __init__(self, usernames: dict[str, UUID | None], tokens: dict[UUID, str]):
+        self.usernames = usernames
+        self.tokens = tokens
+
+    @classmethod
+    async def from_db(cls, db: Database):
+        usernames: dict[str, UUID | None] = {row["username"]: None for row in await db.fetch("SELECT username FROM users;")}
+        tokens = {row["token"]: row["username"] for row in await db.fetch("SELECT username, token FROM tokens;")}
+        for token, username in tokens.items():
+            usernames[username] = token
+        return cls(usernames, tokens)
+
+    def __contains__(self, username: str) -> bool:
+        return username in self.usernames
+
+    @overload
+    def __getitem__(self, key: str) -> UUID | None: ...
+
+    @overload
+    def __getitem__(self, key: UUID) -> str: ...
+
+    def __getitem__(self, key):
+        if isinstance(key, UUID):
+            return self.tokens.get(key, None)
+        return self.usernames.get(key, None)
+
+    def __setitem__(self, key: str, value: UUID | None):
+        self.usernames[key] = value
+        if value:
+            self.tokens[value] = key
+
+    def __delitem__(self, key: str | UUID):
+        if isinstance(key, UUID):
+            self.usernames[self.tokens[key]] = None
+            del self.tokens[key]
+        else:
+            tok = self.usernames[key]
+            if tok:
+                del self.tokens[tok]
+            del self.usernames[key]
+
+def check_username(coro):
+    async def wrapper(self: Database, username: str, *args, **kwargs):
+        if not self.user_exists(username):
+            raise UnknownUsernameError(username)
+        return await coro(self, username, *args, **kwargs)
+    return wrapper
+
 
 class Database:
     pool: asyncpg.Pool
+    cache: UserCache
 
     def __init__(self, app: USACOAPI):
         app.router.lifespan_context = self.lifespan
@@ -97,9 +150,13 @@ class Database:
             if wait:
                 await asyncio.sleep(24 * 60 * 60)
             else: wait = True
-            await self.execute("""
-                DELETE FROM tokens WHERE created < NOW() - INTERVAL '1 month';
+            deleted = await self.fetch("""
+                DELETE FROM tokens
+                    WHERE created < NOW() - INTERVAL '1 month'
+                    RETURNING token;
             """)
+            for row in deleted:
+                del self.cache[row["token"]]
 
     @asynccontextmanager
     async def lifespan(self, app: USACOAPI):
@@ -109,14 +166,30 @@ class Database:
         with open("setup.sql", "r") as f:
             setup = f.read()
         await self.execute(setup)
+        self.cache = await UserCache.from_db(self)
         task = asyncio.create_task(self.invalidate_tokens())
         yield
         # Close the pool
         task.cancel()
         await self.pool.close()
 
+    def user_exists(self, name: str) -> bool:
+        """Checks if a user exists"""
+        return name in self.cache
+
+    def get_username(self, token: UUID) -> str | None:
+        """Gets the username associated with a token"""
+        return self.cache[token]
+
+    def autocomplete(self, name: str) -> list[str]:
+        """Matches partial username with existing usernames"""
+        name = name.lower()
+        return [user for user in self.cache.usernames if user.lower().startswith(name)]
+
     async def signup(self, username: str, password: str, *, conn: OptConn = None) -> str:
         """Register a user into the database and return the token generated"""
+        if self.user_exists(username):
+            raise DuplicateError
         acquired = False
         if not conn:
             acquired = True
@@ -134,13 +207,17 @@ class Database:
             await self.release(conn)
             raise exc from None
         if acquired: await self.release(conn)
+        self.cache[username] = token["token"]
         return str(token["token"])
 
+    @check_username
     async def update_user(
             self, username: str, new_username: str | None = None,
             new_password: str | None = None, *, conn: OptConn = None
     ) -> str:
         """Updates a user's username or password and generate a new token"""
+        if new_username and self.user_exists(new_username):
+            raise DuplicateError
         acquired = False
         if not conn:
             acquired = True
@@ -167,22 +244,20 @@ class Database:
             await self.release(conn)
             raise exc from None
         if acquired: await self.release(conn)
+        del self.cache[username]
+        self.cache[new_username or username] = token["token"]
         return str(token["token"])
 
+    @check_username
     async def remove_user(self, username: str, *, conn: OptConn = None) -> None:
         """Removes a user from the database"""
-        await self.execute(
-            "DELETE FROM users WHERE username = $1;", username, conn=conn
-        )
+        if self.user_exists(username):
+            await self.execute(
+                "DELETE FROM users WHERE username = $1;", username, conn=conn
+            )
+            del self.cache[username]
 
-    async def user_exists(self, name: str, *, conn: OptConn = None) -> bool:
-        """Checks if a user exists"""
-        rec = await self.fetchrow(
-            "SELECT exists (SELECT 1 FROM users WHERE username = $1 LIMIT 1);",
-            name, conn=conn
-        )
-        return rec["exists"]
-
+    @check_username
     async def get_token(self, username: str, password: str, *, conn: OptConn = None) -> dict:
         acquired = False
         if not conn:
@@ -209,6 +284,8 @@ class Database:
             await self.release(conn)
             raise exc from None
         if acquired: await self.release(conn)
+        if not self.cache[username]:
+            self.cache[username] = rec["token"]
         return {
             "token": str(rec["token"]),
             "expiry": str(date(
@@ -218,50 +295,39 @@ class Database:
         }
 
     async def remove_token(self, username: str, *, conn: OptConn = None) -> None:
-        await self.execute(
-            "DELETE FROM tokens WHERE username = $1;", username, conn=conn
-        )
+        if self.cache[username]:
+            await self.execute(
+                "DELETE FROM tokens WHERE username = $1;", username, conn=conn
+            )
+            self.cache[username] = None
 
-    async def get_username(self, token: str, *, conn: OptConn = None) -> str | None:
-        """Gets the username associated with a token"""
-        try:
-            UUID(token)
-        except ValueError:
-            return None
-        rec = await self.fetchrow(
-            "SELECT username FROM tokens WHERE token = $1;", token, conn=conn
-        )
-        if not rec: return None
-        return rec["username"]
-
+    @check_username
     async def get_progress(
-            self, tname: Literal["modules", "problems"], username: str, *, conn: OptConn = None
+            self, username: str, tname: Literal["modules", "problems"], *, conn: OptConn = None
     ) -> dict:
         if tname not in ("modules", "problems"):
+            if conn:
+                await self.release(conn)
             raise ValueError("tname must be either 'modules' or 'problems'")
-        try:
-            data = {
-                row["id"]: dict(row)
-                for row in (await self.fetch(
-                    f"""SELECT id, progress, lastUpdated FROM {tname} WHERE username = $1;""",
-                    username, conn=conn
-                ))
-            }
-            for value in data.values():
-                value["lastUpdated"] = value["lastupdated"].timestamp()
-                del value["lastupdated"]
-            return data
-        except asyncpg.ForeignKeyViolationError:
-            raise UnknownUsernameError(username)
+        data = {
+            row["id"]: dict(row)
+            for row in (await self.fetch(
+                f"""SELECT id, progress, lastUpdated FROM {tname} WHERE username = $1;""",
+                username, conn=conn
+            ))
+        }
+        for value in data.values():
+            value["lastUpdated"] = value["lastupdated"].timestamp()
+            del value["lastupdated"]
+        return data
 
+    @check_username
     async def get_follows(self, username: str, *, conn: OptConn = None) -> list[str]:
-        try:
-            return [row["follow"] for row in (await self.fetch("""
-                SELECT follow FROM follows WHERE username = $1 ORDER BY follow ASC;
-            """, username, conn=conn))]
-        except asyncpg.ForeignKeyViolationError:
-            raise UnknownUsernameError(username)
+        return [row["follow"] for row in (await self.fetch("""
+            SELECT follow FROM follows WHERE username = $1 ORDER BY follow ASC;
+        """, username, conn=conn))]
 
+    @check_username
     async def update_progress(
             self, username: str, tname: Literal["modules", "problems"],
             proglist: ModuleList | ProblemList, *, conn: OptConn = None
@@ -269,23 +335,23 @@ class Database:
         if tname not in ("modules", "problems"):
             raise ValueError("tname must be either 'modules' or 'problems'")
         if len(getattr(proglist, tname)) == 0: return
-        try:
-            await self.executemany(f"""
-                INSERT INTO {tname} (id, username, progress, lastUpdated)
-                    VALUES ($1, $2, $3, $4)
-                    ON CONFLICT (id, username)
-                        DO UPDATE SET progress = $3, lastUpdated = $4;
-            """, [
-                (progress.id, username,
-                 progress.progress, datetime.fromtimestamp(progress.lastUpdated))
-                for progress in getattr(proglist, tname)
-            ], conn=conn)
-        except asyncpg.ForeignKeyViolationError:
-            raise UnknownUsernameError(username)
+        await self.executemany(f"""
+            INSERT INTO {tname} (id, username, progress, lastUpdated)
+                VALUES ($1, $2, $3, $4)
+                ON CONFLICT (id, username)
+                    DO UPDATE SET progress = $3, lastUpdated = $4;
+        """, [
+            (progress.id, username,
+            progress.progress, datetime.fromtimestamp(progress.lastUpdated))
+            for progress in getattr(proglist, tname)
+        ], conn=conn)
 
+    @check_username
     async def follow(
         self, username: str, to_follow: str, *, conn: OptConn = None
     ):
+        if not self.user_exists(to_follow):
+            raise UnknownUsernameError(to_follow)
         acquired = False
         if not conn:
             acquired = True
@@ -297,15 +363,12 @@ class Database:
         except asyncpg.UniqueViolationError:
             await self.release(conn)
             raise DuplicateError
-        except asyncpg.ForeignKeyViolationError:
-            await self.release(conn)
-            raise UnknownUsernameError(username)
         except Exception as exc:
             await self.release(conn)
             raise exc from None
         ret: dict[str, dict] = {}
         for tname in ("problems", "modules"):
-            ret[tname] = await self.get_progress(tname, to_follow, conn=conn) # type: ignore
+            ret[tname] = await self.get_progress(to_follow, tname, conn=conn)
         if acquired:
             await self.release(conn)
         return ret
@@ -313,10 +376,12 @@ class Database:
     async def unfollow(
         self, username: str, to_unfollow: str, *, conn: OptConn = None
     ):
-        await self.execute("""
-            DELETE FROM follows WHERE username = $1 AND follow = $2;
-        """, username, to_unfollow, conn=conn)
+        if self.user_exists(username) and self.user_exists(to_unfollow):
+            await self.execute("""
+                DELETE FROM follows WHERE username = $1 AND follow = $2;
+            """, username, to_unfollow, conn=conn)
 
+    @check_username
     async def get_follower_progress(self, username: str, *, conn: OptConn = None) -> dict:
         acquired = False
         if not conn:
